mruby-bin-debugger: Fix compile without C99 or C11
STRTOUL macro is not support non-C99 and non-C11 mode.
Therefore, if this was compiled without "-std=gnu99" option, we get
the following error.

----
error: $B!F(Bfor$B!G(B loop initial declarations are only allowed in C99 or C11 mode
----

This fixes syntax of STRTOUL macro. And this changes so that it can be compiled
when C99 or C11 mode has not been specified.

Signed-off-by: Nobuhiro Iwamatsu <iwamatsu@nigauri.org>